# Optimize Vector4/Color4 conversions with Unsafe.As