--- a/src/OpenTK.Mathematics/Data/Color4.cs
+++ b/src/OpenTK.Mathematics/Data/Color4.cs
@@ -156,16 +156,11 @@ namespace OpenToolkit.Mathematics
         /// Returns this Color4 as a Vector4. The resulting struct will have XYZW mapped to RGBA, in that order.
         /// </summary>
         /// <param name="c">The Color4 to convert.</param>
-        /// <returns>The Color4 to convert into a Vector4.</returns>
+        /// <returns>The Color4, converted into a Vector4.</returns>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static explicit operator Vector4(Color4 c)
         {
-            unsafe
-            {
-                // Because these structs are identical on the byte-level (due to the StructLayout above),
-                // we can use pointer-casting to convert this into a Vector4, which lets us reinterpret the data instead of copying it to a new struct.
-                return *((Vector4*)&c);
-            }
+            return Unsafe.As<Color4, Vector4>(ref c);
         }
 
         /// <summary>