### Purpose of this PR

* Optimize Vector4/Color4 conversions using the Unsafe.As function.
* Affects OpenTK.Mathematics.

### Testing status

* Compiled it, checked results, that's basically it.

### Comments

Benchmark results, via BenchmarkDotNet:

```
               Method |      Mean |     Error |    StdDev |
--------------------- |----------:|----------:|----------:|
 ToVector4PointerCast | 11.336 ns | 0.2555 ns | 0.2840 ns |
         ToVector4New |  7.740 ns | 0.0446 ns | 0.0396 ns |
    ToVector4UnsafeAs |  2.442 ns | 0.0397 ns | 0.0331 ns |
```

PointerCast is what we were using before (`*((Color4*)&v);`), New is `new Vector4(r, g, b, a)`, and UnsafeAs is the new implementation.