@Nihlus 2019-01-13
Yes, the `Unsafe` library is a .NET Standard library.